Add a central location for managing HLE sockets
