+++ Core/Core.vcxproj.filters
@@ -1342,6 +1342,9 @@
     <ClCompile Include="HLE\NetInetConstants.cpp">
       <Filter>HLE\Libraries</Filter>
     </ClCompile>
+    <ClCompile Include="HLE\SocketManager.cpp">
+      <Filter>HLE\Libraries</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ELF\ElfReader.h">
@@ -2158,6 +2161,9 @@
       <Filter>Net</Filter>
     </ClInclude>
     <ClInclude Include="HLE\NetInetConstants.h">
+      <Filter>HLE\Libraries</Filter>
+    </ClInclude>
+    <ClInclude Include="HLE\SocketManager.h">
       <Filter>HLE\Libraries</Filter>
     </ClInclude>
   </ItemGroup>
