+++ Core/HLE/sceNetInet.h
@@ -186,24 +186,3 @@ int sceNetApctlConnect(int connIndex);
 int sceNetInetPoll(u32 fdsPtr, u32 nfds, int timeout);
 int sceNetApctlTerm();
 int sceNetApctlDisconnect();
-
-enum class SocketState {
-	Unused,
-	Used,
-};
-
-// Internal socket state tracking
-struct InetSocket {
-	SOCKET sock;  // native socket
-	SocketState state;
-	// NOTE: These are the PSP types for now
-	int domain;
-	int type;
-	int protocol;
-	// These are the host types for convenience.
-	int hostDomain;
-	int hostType;
-	int hostProtocol;
-};
-
-extern InetSocket g_inetSockets[256];
